don't check chmod err on windows

+++ diskv.go
@@ -11,6 +11,7 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"runtime"
 	"strings"
 	"sync"
 	"syscall"
@@ -208,7 +209,7 @@ func (d *Diskv) createKeyFileWithLock(pathKey *PathKey) (*os.File, error) {
 			return nil, fmt.Errorf("temp file: %s", err)
 		}
 
-		if err := f.Chmod(d.FilePerm); err != nil {
+		if err := f.Chmod(d.FilePerm); err != nil && runtime.GOOS != "windows" {
 			f.Close()           // error deliberately ignored
 			os.Remove(f.Name()) // error deliberately ignored
 			return nil, fmt.Errorf("chmod: %s", err)
